Implmented User management and changed DB schema

+++ data.js
@@ -9,21 +9,36 @@ var connPool = mysql.createPool({
   password: "6611", 
 });
 
-// Input: data object {title, done, deadline}
-// Output: the id of new Todo
-async function addTodo(data) {
-  const { title, done, deadline } = data;
-  if (title === undefined || done === undefined || deadline === undefined) {
+// Helper functions
+// Input: userName
+// Output: userId
+async function getUserId(userName) {
+  const res = await connPool.awaitQuery("SELECT id FROM Users WHERE username = ?", [userName]);
+  if (res[0]) {
+    return res[0].id; 
+  }
+  else {
     return -1;
   }
-  const res = await connPool.awaitQuery("INSERT INTO Todo (title, done, deadline) VALUES (?, ?, ?);", [title, done, deadline]);
+}
+
+// Input: data object {title, done, deadline, userName}
+// Output: the id of new Todo
+async function addTodo(data) {
+  const { title, done, deadline, userName } = data;
+  const userId = await getUserId(userName);
+  if (title === undefined || done === undefined || deadline === undefined || userId === -1) {
+    return -1;
+  }
+  const res = await connPool.awaitQuery("INSERT INTO Todo (title, done, deadline, userId) VALUES (?, ?, ?, ?);", [title, done, deadline, userId]);
   return res.insertId;
 }
 
-// Input: id (int)
+// Input: id (int), userName (str)
 // Output: true if deleted, false otherwise
-async function deleteTodo(id) {
-  const res = await connPool.awaitQuery("DELETE FROM Todo WHERE id = ?", [id]);
+async function deleteTodo(id, userName) {
+  const userId = await getUserId(userName);
+  const res = await connPool.awaitQuery("DELETE FROM Todo WHERE id = ? AND userId = ?", [id, userId]);
   if (res.affectedRows === 0) {
     return false;
   }
@@ -32,40 +47,74 @@ async function deleteTodo(id) {
 
 // Input: option (string) one of the following (overdue, done, undone, all)
 // Output: list of matching Todos ordered by done status (if applicable) and due date 
-async function getTodos(option) {
+async function getTodos(option, userName) {
+  const userId = await getUserId(userName);
   let todos = [];
   if (option == "overdue") {
-    todos = await connPool.awaitQuery("SELECT * FROM Todo WHERE deadline < CURDATE() AND done = false ORDER BY deadline");
+    todos = await connPool.awaitQuery("SELECT id, title, done, deadline FROM Todo WHERE deadline < CURDATE() AND done = false AND userId = ? ORDER BY deadline", [userId]);
   }
   else if (option == "done") {
-    todos = await connPool.awaitQuery("SELECT * FROM Todo WHERE done = true ORDER BY deadline"); // is it true or 1
+    todos = await connPool.awaitQuery("SELECT id, title, done, deadline FROM Todo WHERE done = true AND userId = ? ORDER BY deadline", [userId]); 
   }
   else if (option == "undone") {
-    todos = await connPool.awaitQuery("SELECT * FROM Todo WHERE done = false ORDER BY deadline"); // is it false or 0
+    todos = await connPool.awaitQuery("SELECT id, title, done, deadline FROM Todo WHERE done = false AND userId = ? ORDER BY deadline", [userId]); 
   }
   else {
-    todos = await connPool.awaitQuery("SELECT * FROM Todo ORDER BY done, deadline");
+    todos = await connPool.awaitQuery("SELECT id, title, done, deadline FROM Todo WHERE userId = ? ORDER BY done, deadline", [userId]);
   }
   return todos;
 }
 
-// Input: data object {id, title, done, deadline}
+// Input: data object {id, title, done, deadline, userName}
 // Output: true if updated, false otherwise
 async function updateTodo(data) {
-  let { id, title, done, deadline } = data;
-  console.log(title, done, deadline)
-  const res = await connPool.awaitQuery("UPDATE Todo SET title = ?, done = ?, deadline = ? WHERE id = ?", [title, done, deadline, id]);
+  let { id, title, done, deadline, userName } = data;
+  const userId = await getUserId(userName);
+  const res = await connPool.awaitQuery("UPDATE Todo SET title = ?, done = ?, deadline = ? WHERE id = ? AND userId = ?", [title, done, deadline, id, userId]);
   if (res.affectedRows === 0) {
     return false;
   }
   return true;
 }
 
-// Input: id (int)
-// Output: matching todo
-async function getTodoById(id) {
-  const todo = await connPool.awaitQuery("SELECT * from Todo WHERE id = ?", [id]);
+// Input: id (int), userName (str)
+// Output: matching todo or empty list
+async function getTodoById(id, userName) {
+  const userId = await getUserId(userName);
+  const todo = await connPool.awaitQuery("SELECT * from Todo WHERE id = ? AND userId = ?", [id, userId]);
   return todo;
+}
+
+// Input: data {username, password}
+// Output: User id if added, -1 otherwise
+async function addUser(data) {
+  let {username, password} = data;
+  const user = await connPool.awaitQuery("SELECT * from Users WHERE username = ?", [username]);
+  if (user.length !== 0) {
+    return -1;
+  }
+  const res = await connPool.awaitQuery("INSERT INTO Users (username, password) VALUES (?, ?)", [username, password]);
+  return res.insertId;
+}
+
+// Input: userName
+// Output: password if user exists, empty string otherwise
+async function getPassword(userName) {
+  const res = await connPool.awaitQuery("SELECT password FROM Users WHERE username = ?", [userName]);
+  if (res.length === 0) {
+    return "";
+  }
+  return res[0].password;
+}
+
+// Input: todoId
+// Output: true if done state changed, false otherwise
+async function changeDoneState(todoId) {
+  const res = await connPool.awaitQuery("UPDATE Todo SET done = !done WHERE id = ?", [todoId]);
+  if (res.affectedRows === 0) {
+    return false;
+  }
+  return true;
 }
 
 module.exports = {
@@ -73,5 +122,9 @@ module.exports = {
   deleteTodo,
   getTodos,
   updateTodo,
-  getTodoById
+  getTodoById,
+  addUser,
+  getPassword,
+  getUserId, 
+  changeDoneState
 };
